Fixed POST binding issue

+++ Backend/Models/Product.cs
@@ -3,7 +3,8 @@
     public class Product
 {
     public int Id { get; set; }
-    public string Name { get; set; }
-    public double Price { get; set; }
+
+    public string Name { get; set; }   // ✅ NOT NULL
+    public decimal Price { get; set; } // ✅ NOT NULL
 }
 }
